Added Interval, Hittable List, refactor color calculation in software renderer into a function using hittable list.

+++ src/Private/Color.cpp
@@ -7,9 +7,9 @@ void WriteColor(std::ostream& OutFileStream, const Color& PixelColor)
 	float B = PixelColor.B();
 
 	//In case we want to run to a ppm file, not needed for direct render to screen
-	int AdjustedRed = 255.999f * R;
-	int AdjustedGreen = 255.999f * G;
-	int AdjustedBlue = 255.999f * B;
+	int AdjustedRed = (int)(255.999f * R);
+	int AdjustedGreen = (int)(255.999f * G);
+	int AdjustedBlue = (int)(255.999f * B);
 
 	OutFileStream << AdjustedRed << ' ' << AdjustedGreen << ' ' << AdjustedBlue << '\n';
 }
